+++ src/refactoring/components/Admin/product/ProductDiscount/ProductDiscountUpdate.tsx
@@ -9,8 +9,6 @@ interface Props {
 }
 
 export const ProductDiscountUpdate = ({ product }: Props) => {
-  const [updatedProduct, setUpdatedProduct] = useState<Product>(product);
-
   const { editingProduct, setEditingProduct, updateProduct } =
     useProductStore();
 
@@ -40,21 +38,6 @@ export const ProductDiscountUpdate = ({ product }: Props) => {
     setNewDiscount({ quantity: 0, rate: 0 }); // 새로운 할인 정보 초기화
   };
 
-  // 할인 정보 삭제
-  const handleRemoveDiscount = (index: number) => {
-    const updatedProduct = {
-      ...product,
-      discounts: product.discounts.filter((_, i) => i !== index),
-    };
-
-    console.log("updatedProduct", updatedProduct);
-
-    updateProduct(updatedProduct);
-
-    // 상태 업데이트 후, 바로 DOM에 반영되도록 처리
-    setEditingProduct(updatedProduct);
-  };
-
   return (
     <>
       <div>
@@ -62,9 +45,8 @@ export const ProductDiscountUpdate = ({ product }: Props) => {
         {product.discounts.map((discount, index) => (
           <ProductDiscountInfoItem
             key={index}
-            discount={discount}
-            handleRemoveDiscount={handleRemoveDiscount}
             product={product}
+            discount={discount}
             index={index}
           />
         ))}
